Tidied something up
LevelInfoDisplay was passing itself as a GameObject to PlayerControl,
which was then doing GetComponent<LevelInfoDisplay> on the Go to get it.

This was insane, so I have stopped it.

(Also in last commit, I change the PlayerPref names for everything but
WebGL. This meant I also did a wipe of the old ones before as well).;