Fixed some stuff
First up, I guess I added PostProcessing stuff to the project

Secondly, fixed some bugs with the previous commit:

Changed variable names to better convey their purpose
Changed levelStats to fill out from the complete list of levels, not
just the ones in the 'queue'
Changed PlayerPrefs to update from level ID, not the queuePosition from
range.

More notes are in Trello on this, as the above barely makes sense.